Change array concat to string concat.
In recent versions of CPython, string += is said to be optimized to avoid the O(n^2)
behavior which leads to the idiom of building up things as arrays.

Another alternative would be to make an array of strings and then join them.

Or just not worry about it; I'm not sure the strings here get big enough to matter.